requirements.txt
discord.py>=2.3.0
ollama>=0.4.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
trafilatura>=1.6.0
ddgs>=7.0.0
